Rework everything relating to icons.
Since our first attempt, Google decided to totally rework adaptive
icons. This commit fixes the shit that was broken from the original
commit. we will mwerge them together before we push to any other rom.
(if we did this originally, we could have pushed it to other roms months
back)